# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de registro de asistencia",
        "description": "listado de URI's de la API alumnos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/alumnos": {
            "get": {
                "tags": [
                    "Alumno"
                ],
                "summary": "lista de todos los alumnos",
                "operationId": "52ca0d8dad6b04f9c7e3f43eca221f9d",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Paolo Lucio"
                                                    },
                                                    "apellido": {
                                                        "type": "string",
                                                        "example": "Conza HAquehua"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "paolo@gmail.com"
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "example": "987288269"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Registrar a un nuevo Alumno",
                "operationId": "e9e77b336781c9f8e41e532fef3fb69d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Paolo Lucio",
                                    "apellido": "Conza Haquehua",
                                    "email": "paolo@gmail.com",
                                    "telefono": "12345678"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Paolo Lucio"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Conza Haquehua"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "paolo@gmail.com"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "12345678"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alumnos/{id}": {
            "put": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Actualizar la información de un alumno",
                "operationId": "bfa9cf00f2ef20fc3e4a526ba0ca66a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Paolo Lucio Editado",
                                    "apellido": "Conza Haquehua Editado",
                                    "email": "paolo@gmail.com Editado",
                                    "telefono": "12345678 Editado"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Paolo Lucio Editado"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Conza Haquehua Editado"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "paolo@gmail.com Editado"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "12345678 Editado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Eliminar la información de un alumno",
                "operationId": "6c6ef9d7128996dfd0ee5c79ba9e4bbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}